pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["viacheslavzablotskiy <zlava.mag@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.104.0"}
pydantic = "^2.4.2"
uvicorn = "^0.23.2"
asyncio = "^3.4.3"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.8.0"
sqladmin = "^0.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
